==> 01_extraction/01-datasets_overview.ipynb <==
Let's see an overview of the datasets

# Necessary imports to read the datasets
import pandas as pd
import zipfile
import os
import numpy as np


# Checking environment
import sys
print(sys.executable)

# DAILY DATA
## First dataset
Within the daily_consumption dataset we have data divided in two csv files. Starting with the daily_consumption:

df_daily_census = pd.read_csv("../data/datasets/daily_consumption_census.csv", low_memory=False)


print("DATA AGGREGATED BY CENSUS SECTION:")
df_daily_census

# These are the columns of the dataset
columns_to_check = [
    "Census section",
    "District",
    "Municipality",
    "Date",
    "Use",
    "Number of meters",
    "Accumulated consumption (L/day)"
]

# We will go thorugh all columns to check the different values they get
for col in columns_to_check:
    values = df_daily_census[col].replace('<NULL>', pd.NA).dropna().unique()    # We have to remove the "<NULL>" values
    print(f"COLUMN: {col}") 
    print(f"Unique values: {len(values)}")
    print(f"Values: {values}\n")


## Second dataset
About industrial and comercial use

df_daily_economic = pd.read_csv("../data/datasets/daily_consumption_economic.csv", low_memory=False)

df_daily_economic

# These are the columns of the dataset
columns_to_check = [
    "District",
    "Municipality",
    "Date",
    "Use",
    "Economic activity",
    "Economic activity description",
    "Number of meters",
    "Accumulated consumption (L/day)"
]

# We will go thorugh all columns to check the different values they get
for col in columns_to_check:
    values = df_daily_economic[col].replace('<NULL>', pd.NA).dropna().unique()    # We have to remove the "<NULL>" values
    print(f"COLUMN: {col}") 
    print(f"Unique values: {len(values)}")
    print(f"Values: {values}\n")

# HOURLY DATA
## Dataset
Let's check the hourly consumption data:

df_hourly_consumption = pd.read_parquet("../data/datasets/hourly_consumption.parquet", engine="pyarrow")

# These are the columns of the dataset
columns_to_check = [
    "Policy",
    "Technology",
    "Counter diameter (cm)",
    "Use",
    "Type of housing",
    "Date",
    "Reading index (L/h)"
]

# We will go thorugh all columns to check the different values they get
for col in columns_to_check:
    values = df_hourly_consumption[col].replace('<NULL>', pd.NA).dropna().unique()    # We have to remove the "<NULL>" values
    print(f"COLUMN: {col}") 
    print(f"Unique values: {len(values)}")
    print(f"Values: {values}\n")
